clustermesh-apiserver: use distroless/static-debian11 as the base image
[ upstream commit 1d5615771b3dab1d5eeb9fda18ae0d98a22de73d ]

[ backporter's notes: replaced the nonroot base image with the root one,
  to avoid requiring the Helm changes to configure the fsGroup, which
  could cause issues if users only updated the image version, without
  a full helm upgrade. ]

gops needs to write data (e.g., the PID file) to the file-system, which
turned out to be tricky when using scratch as base image, in case the
container is then run using a non-root UID.

Let's use the most basic version of a distroless image instead, which
contains:

- ca-certificates
- A /etc/passwd entry for a root, nonroot and nobody users
- A /tmp directory
- tzdata

This aligns the clustermesh-apiserver image with the Hubble Relay one,
and removes the need for manually importing the CA certificates. The
GOPS_CONFIG_DIR is explicitly configured to use a temporary directory,
to prevent permission issues depending on the UID configured to run
the entrypoint.

Finally, we explicitly configure the fsGroup as part of the
podSecurityContext, to ensure that mounted files are accessible
by the non-root user as well.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>