workflows: Cover KPR in end-to-end IPsec tests
We can reuse the two configs that had --devices set because KPR will
cause devices to be autodetected anyway. We then need to add one other
config to cover VXLAN.

Upgrade tests are not extended to cover KPR because it isn't supported
in the previous stable. We will need to wait for the next minor release
to be able to extend those tests.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>